MPS doesn't support float64

+++ attributor/attributor.py
@@ -22,7 +22,7 @@ class Attributor:
         """
 
         # Remove batch dimension (which must be one), now A is [heads, len, len]
-        A = attention.squeeze(0).type(torch.float64)
+        A = attention.squeeze(0).type(torch.float32)
         # Sum over attention heads (total attention from token i to j). now A is [len, len]
         A = A.sum(axis=0)
 
@@ -41,7 +41,7 @@ class Attributor:
     def attribute(self, attentions):
         n = attentions[0].shape[2]
         # Values will get close to 0, use lots of precision
-        Y = torch.eye(n, n, dtype=torch.float64).to(attentions[0].device)
+        Y = torch.eye(n, n, dtype=torch.float32).to(attentions[0].device)
         for A in attentions:
             Y = self.forward(Y, A)
